pyproject.toml
[project]
name = "grpc_agents"
version = "0.1.0"
description = "A gRPC service to chat with LLMs and agents."
requires-python = ">=3.12,<3.13"

dependencies = [
    "click",
    "pytest",
    "pytest-asyncio",
    "python-dotenv",
    "PyYAML",
    "ruff",
    "grpcio",
    "grpcio-tools",
    "openai",
    "tiktoken",
    "aiosqlite",
    "anthropic>=0.42.0",
    "aiofiles>=24.1.0",
    "cachetools>=5.5.0",
    "pypdf>=5.1.0",
    "aiohttp>=3.11.11",
    "pathspec>=0.12.1",
    "beautifulsoup4>=4.12.3",
    "pytest-timeout>=2.3.1",
    "sentence-transformers>=3.3.1",
]

[tool.pytest.ini_options]
timeout = 10
timeout_method = "signal"  # note this only works on unix; "thread" method (default) is safer but might not catch hanging subprocesses
pythonpath = [
    "proto",
    "proto/generated"
]
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    # getting a lot of warnings from protobuf
    # /Users/shanekercheval/repos/grpc-agents/.venv/lib/python3.13/site-packages/google/protobuf/internal/well_known_types.py:178: DeprecationWarning: datetime.datetime.utcnow() is deprecated and scheduled for removal in a future version. Use timezone-aware objects to represent datetimes in UTC: datetime.datetime.now(datetime.UTC).
    #     self.FromDatetime(datetime.datetime.utcnow())
    "ignore::DeprecationWarning:google.protobuf.*"
]
